Adds `.keep_next` paragraph property.
> Specifies that the paragraph (or at least part of it) should be
rendered on the same page as the next paragraph when possible. It is an
empty element: <w:keepNext/>. If multiple paragraphs are to be kept
together but they exceed a page, then the set of paragraphs begin on a
new page and page breaks are used thereafter as needed.

http://officeopenxml.com/WPparagraphProperties.php

Usage

```
context.p 'hi' do
  keep_next true
end
```